+++ test/fixtures/taxonomy.py
@@ -72,6 +72,6 @@ def mock_taxonomy_lf() -> dy.LazyFrame[TaxonomySchema]:
         pl.col("family").cast(pl.Categorical),
         pl.col("genus").cast(pl.Categorical),
         pl.col("taxonRank").cast(pl.Enum(TAXON_RANK_VALUES)),
-        pl.col("gbifTaxonId").cast(pl.Int64),
+        pl.col("gbifTaxonId").cast(pl.UInt32),
     )
     return TaxonomySchema.validate(taxonomy_df).lazy()
