+++ test/test_cluster_taxa_heatmap.py
@@ -178,7 +178,7 @@ class TestClusterTaxaHeatmap(unittest.TestCase):
             pl.col("family").cast(pl.Categorical),
             pl.col("genus").cast(pl.Categorical),
             pl.col("taxonRank").cast(pl.Enum(TAXON_RANK_VALUES)),
-            pl.col("gbifTaxonId").cast(pl.Int64),
+            pl.col("gbifTaxonId").cast(pl.UInt32),
         )
         taxonomy_df = TaxonomySchema.validate(taxonomy_df)
 
@@ -350,7 +350,7 @@ class TestClusterTaxaHeatmap(unittest.TestCase):
             pl.col("family").cast(pl.Categorical),
             pl.col("genus").cast(pl.Categorical),
             pl.col("taxonRank").cast(pl.Enum(TAXON_RANK_VALUES)),
-            pl.col("gbifTaxonId").cast(pl.Int64),
+            pl.col("gbifTaxonId").cast(pl.UInt32),
         )
         taxonomy_df = TaxonomySchema.validate(taxonomy_df)
